combine functions so only one inverse done during testing
